langchain_agent/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
python-dotenv>=1.0.0
requests>=2.31.0 